Fix specialty creation and rewrite README
- Co-locate NuccTaxonomyCode with StaffSpecialty to avoid cross-file
  CustomModel FK class-identity mismatch on plugin reload
- Add to_field="dbid" to the nucc_code FK
- Flatten the specialty API response so the JS finds display_name/code
- Rewrite README to cover features, requirements, install, config,
  permissions, custom data namespace, full API reference, tests, license
- Add app screenshot at docs/screenshot.png
- Bump plugin_version to 0.1.12
